Model microfloat weights as quantization
